feat: 🎸 add responsiveness to counter
